create_disk.py: Allow for the system image to be optionally included
Since the system image is so large, allow for the system image to
be optinally included in the disk image. This allows for a smaller
compressed disk image that just includes the boot partition and
empty cache/userdata partitions. The raw system.img can then be
compressed and uploaded separately, and added to the VM as an extra
disk.

Using this method makes it much quicker to modify just the kernel
and upload a much smaller image, and provision a new vm for development
testing.

Signed-off-by: John Stultz <john.stultz@linaro.org>