create_disk: Fix assumption that kernel/ramdisk will have specific names
Marissa Wall noted that the kernel and ramdisk names are
hard coded in the syslinux config, and the script doesn't
correct for this when copying the specified files to the
partition. So if the specified kernel/ramdisk.img are named
differently, the resulting image won't boot.

Fix this by renaming the files as we copy them over.

Reported-by: Marissa Wall <marissaw@google.com>
Signed-off-by: John Stultz <john.stultz@linaro.org>